requirements.txt
tweepy>=4.15.0
python-dotenv>=1.0.0
requests>=2.28.0
feedparser>=6.0.0
